requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.2
celery==5.3.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
dj-database-url==3.0.1
Django==4.2.26
django-celery-beat==2.5.0
django-cors-headers==4.7.0
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.6.0
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.52
PyJWT==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.9.0
